make App.apks List into a single Apk instance at App.installedApk
This then serves to represent the APK that is installed for this app. It
needs to be filled out by the Providers then also.  This then becomes a
place for data specific to the an installed App, like installedVersionCode
and installedVersionName, instead of having it both stored in an App
instance and a related Apk instance.