catch NameNotFoundException when looking up installer name
If an application was installed via ADB, then it won't have an installer
package/name set.  Therefore, looking up the label will cause a
NameNotFoundException, which should be ignored and not passed to the method
that called App() since it does not refer to the packageName that the
method is trying to look up, but instead the packageName of the missing
installer app.