live filtering of InstalledApps for setting up Local Repo
This implements live filtering in a SearchView so that it is easy to search
for the apps you want to include in your Local Repo.  This requires some
newer stuff, so I switched it to the android-11 Activity until appcompat-v7
is included.  All this functionality will work fine with appcompat-v7.