do not include URL in index.xml because it will get out of date
As far as I can tell, the 'url' metadata in index.xml is not used at all by
the client.  In order to keep it up-to-date in the local repo, it would
have to regenerate index.xml and index.jar each time the IP address
changed.  That would mean a decent amount of work happening in the
background, all the update an unused field in index.xml.