include script for jenkins to run for setting up the ant build
Jenkins will then call ant directly after running this script.  This then sets
the debug builds made by Jenkins to have a versionCode of the UNIX time in
seconds and adds the data to the versionName.  This clearly marks these as
debug builds and also makes it possible to have an fdroid repo of debug builds
that will automatically update after each build.