ManageReposActivity: move as much out of the Fragment; use appcompat menu
Now that the Fragment is embedded in the Activity, and the menu has been
moved to the Activity in ActionBar style, most of the utility functions and
the Dialogs can be based out of the Activity, which is how they are
designed to work.  This makes things work a lot easier.

fixes #3 https://gitlab.com/fdroid/fdroidclient/issues/3