split out new, incoming repo configs into a separate class: NewRepoConfig
The swap stuff will also need to handle incoming Intents that represent
new repos, so the parsing logic is now its own class NewRepoConfig, which
is something like the Repo class, but using getters instead of properties.
Since the new repo data does not change once FDroid receives it, the only
way to set the values of a NewRepoConfig is via the constructor.

This is based on some incomplete work from @pserwylo:
https://gitlab.com/pserwylo/fdroidclient/commit/71cb12ef5c2373f2f9247630dccea15d027f0c5b#diff-6
and
https://gitlab.com/pserwylo/fdroidclient/commit/71cb12ef5c2373f2f9247630dccea15d027f0c5b#diff-7