sanitize URL from clipboard and parse fingerprint (fixes #50)
Instead of just sticking whatever URL is in the clipboard into the "Add
Repo" dialog, this attempts to sanitize the URL in case it has some garbage
or came from a QR Code, and therefore was all uppercase (that makes for
smaller QR Codes).  It also checks if there is a fingerprint in the query
string of the URL, and sticks that into the fingerprint box.

fixes #50 https://gitlab.com/fdroid/fdroidclient/issues/50