change title to "F-Droid" for VIEW Intent receiver chooser
When someone clicks on a URL that F-Droid can accept, i.e. a repo URL, then
Android puts up a chooser where the user can select which app to VIEW the
URL with.  That was showing up with the title "Repositories", which is the
title used for that Activity when viewing it.  This keeps the Activity
title the same while changing the title in the chooser.