include relevant sites in list of HTTPS pins
The current HTTPS trust model is to first check if a site's key is TOFUed,
then check if it is pinned and check the CA, then prompt the user.  There
is currently no way to only check the CA for validity.  Ultimately, that
should probably not be needed if the repo URLs can include the HTTPS pin
info in the same way that the repo fingerprint is specified.  Then it can
be added to the TOFU/POP keystore when the user accepts the Add Repo dialog

Since that idea does not exist yet, this commit adds the sites that are
likely to run their own repos in the near future:

https://f-droid.org
https://guardianproject.info
https://s3.amazonaws.com  # multiple orgs use this
https://panicbutton.io    # Amnesty International's app
https://psiphon.ca        # circumvention tool