fix crash after MemorizingTrustManager's TOFU/POP Always/Once/Abort prompt
If there is an unknown HTTPS certificate, MemorizingTrustManager puts up a
prompt to ask whether the user wants to trust the certificate.  It comes at
a weird time in the lifecycle of the dialogs, so the previous dialog might
be null.  Therefore add a null check.

This situation should probably be improved and better integrated.