darken category menu button on the dark theme to match the theme
This commit uses alpha to make the category menu button appear darker to
match the rest of the dark theme.  Since the background is black, the alpha
makes it darker.  It is only used on the dark theme since alpha would
lighten the menu button on the light themes, and that would make it a worse
match.