Prevent crash when changing categories. Fixes #72.
The problem arose when we start with no categories other than the
three defaults, then add a repo with multiple categories. The exact
issue was that although the category spinner itself was updated,
the listener for onChange was referencing the list of categories
from before (with only three categories). Changed it to use data
from the category spinner adapter, that way it is always up to
date.

Also fixed some warnings in the file. Instance access of a static
object, and deprecated method warning.