Prevent android.support.v7.SearchView from being excluded by proguard.
There is a bug where the android.support.widget.SearchView gets
removed by proguard which causes breakage. This prevents that
from occuring by a bit of a brute force approach. While it would
be possible to keep only the class in question, this may leave
us open to other bugs in the future. Better to be safe than sorry.

http://stackoverflow.com/questions/22136032/action-bar-search-view-android-null-error