do not run proguard in `ant debug` builds, it breaks the tests
The whole ant build system was built around the assumption that proguard is
only run on the release builds.  Running proguard on the debug builds
breaks the test instrumentation, which does something like insert the emma
methods into each class.  In theory, this could be fixed, but I think its a
bad idea to heavily customize the ant build system since its basically
deprecated in favor of gradle.

Here's the proguard warning message:
 [proguard] Warning: can't find referenced class com.vladium.emma.rt.RT