Update README after F-Droid file move.
The README documentation was not updated after the move of the f-droid client
code to the F-Droid directory. The instruction where therefore wrong

The troubleshooting documentation about determining the used Android versions
in the project was also outdated. While the script was made functional the tip
to install the various sdk from the command line did does not work for me.